Fixing most of the errors
- Sending data instead of pointers in scatter (sending a pointer to a
  memory in processor x to processor y led to unknown behaviour) by
  creating 1D dynamic array to hold all images in sequential way

- Presenting global variables need by MPI

- the output is not correct and needs to be fixed

+++ HPC_ProjectTemplate/Source.cpp
@@ -14,12 +14,16 @@
 using namespace std;
 using namespace msclr::interop;
 
+// Image Global Macros
+#define IMAGE_NUMBER 495
+#define IMAGE_HEIGHT 320
+#define IMAGE_WIDTH 240
+#define IMAGE_SIZE 76800
 
-// Image Global Variables
-int const IMAGE_NUMBER = 40;
-int const IMAGE_HEIGHT = 320;
-int const IMAGE_WIDTH = 240;
-int const IMAGE_SIZE = 76800;
+// MPI Global Variables
+int wSize, wRank;
+const int rootProcessor = 0;
+
 
 // Load image to memory
 int* inputImage(int* w, int* h, System::String^ imagePath);
@@ -28,51 +32,77 @@ int* inputImage(int* w, int* h, System::String^ imagePath);
 void createImage(int* image, int width, int height, int index);
 
 // Load multiple images
-void getImages(int**& imgArr);
+void getImages(int** imgArr);
+
+// Copy Image from source to destination
+void copyImage(int* dst, int* src);
 
 // Create an array of images in the heap
 void initializeImageArray(int**& imgArr, int const sz);
 
-// Free the heap from an array of images
-void deleteImageArray(int**& imgArr, int const sz);
+// Free the heap from an array of images - the array must be declared using initializeImageArray()
+void deleteImageArray(int** imgArr, int const sz);
 
 // Add Image Array
-void addImageArray(void* inputBuffer, void* outputBuffer, int* len, MPI_Datatype* datatype);
-
+void addImageArray(int** inputBuffer, int* outputBuffer, int len);
 
 // Get the mean of the pixels
-void pixelMean(int* &img);
+void pixelMean(int* img);
 
 // Get the background image
-int* backgroundExtraction(int** &imgArr);
+int* backgroundExtraction(int** imgArr);
 
 int main()
 {
+	// According to this answer https://stackoverflow.com/questions/2156714/how-to-speed-up-this-problem-by-mpi/2290951#2290951
+	// All the code works in parallel not only MPI_Init() & MPI_Finalize() part
+
+	// Initialize the MPI env
+	MPI_Init(NULL, NULL);
+
+	// Get processes number
+	MPI_Comm_size(MPI_COMM_WORLD, &wSize);
+	// Get processor rank
+	MPI_Comm_rank(MPI_COMM_WORLD, &wRank);
+
 	//Create array of images
-	int** imgArr = new int* [IMAGE_NUMBER] {0};
-	
-	// Fill the array with images
-	getImages(imgArr);
+	int** imgArr = nullptr;
+	initializeImageArray(imgArr, IMAGE_NUMBER);
+
+	// Sequential part
+	if (wRank == rootProcessor) {
+		// Fill the array with images
+		getImages(imgArr);
+	}
 
 	// Start Parrallel code
 	int start_s = clock();
 
 	// Get background image
-	int *bgImg = backgroundExtraction(imgArr);
+	int* bgImg = backgroundExtraction(imgArr);
 
 	// End Parrallel code
 	int stop_s = clock();
 
 	// Get Parrallelism time
 	int TotalTime = (stop_s - start_s) / double(CLOCKS_PER_SEC) * 1000;
-	cout << "time: " << TotalTime << endl;
 
-	// Save background image
-	createImage(bgImg, IMAGE_WIDTH, IMAGE_HEIGHT, 0);
+	// Sequential part
+	if (wRank == rootProcessor) {
+		// print Rank 0 time only
+		cout << "time: " << TotalTime << endl;
 
-	//Free memory
-	deleteImageArray(imgArr, IMAGE_NUMBER);
-	delete[] bgImg;
+		// Save background image
+		createImage(bgImg, IMAGE_WIDTH, IMAGE_HEIGHT, 1);
+
+		//Free memory
+		deleteImageArray(imgArr, IMAGE_NUMBER);
+		// Only rank 0 will return a value
+		delete[] bgImg; 
+	}
+
+	// Finalize MPI env
+	MPI_Finalize();
 
 	return 0;
 }
@@ -144,14 +174,14 @@ void createImage(int* image, int width, int height, int index)
 }
 
 
-void getImages(int** &imgArr) {
+void getImages(int** imgArr) {
 	System::String^ imagePath;
 	string const path = "../Data/Input/";
 	string imgName;
 	string num;
 	int internalLoop;
 
-	int ImageWidth, ImageHeight, const totalDigits = 6;
+	int *img, ImageWidth, ImageHeight, const totalDigits = 6;
 
 	for (int i = 1; i <= IMAGE_NUMBER; i++) {
 		// Get Img Name
@@ -167,103 +197,92 @@ void getImages(int** &imgArr) {
 
 		// Get Img
 		imagePath = marshal_as<System::String^>(path + imgName);
-		imgArr[i - 1] = inputImage(&ImageWidth, &ImageHeight, imagePath);
+		img = inputImage(&ImageWidth, &ImageHeight, imagePath);
+
+		copyImage(imgArr[i - 1], img);
+		
+		delete[] img;
 	}
 }
 
 
-void initializeImageArray(int** &imgArr, int const sz) {
+void copyImage(int* dst, int* src) {
+	for (int i = 0; i < IMAGE_SIZE; i++)
+		dst[i] = src[i];
+}
+
+
+void initializeImageArray(int**& imgArr, int const sz) {
+	// Initialize contiguous dynamic array
+	int* tmp = new int[sz * IMAGE_SIZE]{ 0 };
+
+	// Convert it to 2D array
 	imgArr = new int* [sz];
 	for (int i = 0; i < sz; i++)
-		imgArr[i] = new int[IMAGE_SIZE] {0};
+		imgArr[i] = tmp + (i * IMAGE_SIZE);
 }
 
 
-void deleteImageArray(int** &imgArr, int const sz) {
-	for (int i = 0; i < sz; i++)
-		delete[] imgArr[i];
-	delete[] imgArr;
+void deleteImageArray(int** imgArr, int const sz) {
+	delete[] * imgArr;
 }
 
 
-void addImageArray(void* inputBuffer, void* outputBuffer, int* len, MPI_Datatype* datatype) {
-	// convert them into a proper datatype
-	int** imgArr = (int**)inputBuffer;
-	int* newImg = (int*)outputBuffer;
+void addImageArray(int** inputBuffer, int* outputBuffer, int len) {
+	if (len == 0)
+		return;
+
 	int sum = 0;
-
-	// int* len -> is the "count" argument passed to the reduction call.
-	// or if called by user it's the array size
-
 	for (int i = 0; i < IMAGE_SIZE; i++) {
-		for (int j = 0; i < (*len); j++) {
-			sum += imgArr[j][i];
+		for (int j = 0; j < len; j++) {
+			sum += inputBuffer[j][i];
 		}
-		newImg[i] += sum;
+		outputBuffer[i] += sum;
 		sum = 0;
 	}
 }
 
 
-void pixelMean(int* &img) {
+void pixelMean(int* img) {
 	for (int i = 0; i < IMAGE_SIZE; i++)
 		img[i] /= IMAGE_NUMBER;
 }
 
 
-int* backgroundExtraction(int** &imgArr) {
-	int sz;
+int* backgroundExtraction(int** imgArr) {
+	// Final Image
 	int* retImg = new int[IMAGE_SIZE] {0};
 
-	// Initialize the MPI env
-	MPI_Init(NULL, NULL);
-
-	// Get processes number
-	int wSize;
-	MPI_Comm_size(MPI_COMM_WORLD, &wSize);
-
-	// Create Image Type in MPI
-	MPI_Datatype MPI_IMAGE;
-	MPI_Type_contiguous(IMAGE_SIZE, MPI_INT, &MPI_IMAGE);
-	MPI_Type_commit(&MPI_IMAGE);
-
-	// Create Add Image Operation in MPI
-	MPI_Op MPI_ADD_IMAGE;
-	MPI_Op_create((MPI_User_function*)addImageArray, 1, &MPI_ADD_IMAGE);
-
 	// divide imgs on processors
-	sz = IMAGE_NUMBER / wSize;
+	int sz = IMAGE_NUMBER / wSize;
+
+	// Create a buffer for each processor
 	int** buffer = nullptr;
 	initializeImageArray(buffer, sz);
 
-	cout << "Ready for Scatter .. Press any key\n";
-	cin.get();
-
-	MPI_Scatter(imgArr, sz, MPI_IMAGE, buffer, sz, MPI_IMAGE, 0, MPI_COMM_WORLD);
-
-	cout << "Done Scatter .. Press any key\n";
-	cin.get();
+	// Scatter the data on the processors
+	MPI_Scatter(*imgArr, sz * IMAGE_SIZE, MPI_INT, *buffer, sz * IMAGE_SIZE, MPI_INT, rootProcessor, MPI_COMM_WORLD);
 
 	// put the result from each processor
 	int* sendImg = new int[IMAGE_SIZE] {0};
-	addImageArray(buffer, sendImg, &sz, nullptr);
+	// Add the pixels
+	addImageArray(buffer, sendImg, sz);
 
 	// reduce the result at root processor = 0
-	MPI_Reduce(sendImg, retImg, 1, MPI_IMAGE, MPI_ADD_IMAGE, 0, MPI_COMM_WORLD);
+	MPI_Reduce(sendImg, retImg, IMAGE_SIZE, MPI_INT, MPI_SUM, rootProcessor, MPI_COMM_WORLD);
 
-	int wRank;
-	MPI_Comm_rank(MPI_COMM_WORLD, &wRank);
-	if (wRank == 0) {
+	// Sequential part
+	if (wRank == rootProcessor) {
 		int* remainedImgs = new int[IMAGE_SIZE] {0};
 
 		// Add remained images
 		int start = wSize * (IMAGE_NUMBER / wSize);
-		int imgNum = IMAGE_NUMBER;
-		addImageArray(imgArr + start, remainedImgs, &imgNum, nullptr);
+		int imgNum = IMAGE_NUMBER - start;
+		addImageArray(imgArr + start, remainedImgs, imgNum);
 
 		// Add last 2 img
 		imgNum = 1;
-		addImageArray(remainedImgs, retImg, &imgNum, nullptr);
+		addImageArray(&remainedImgs, retImg, imgNum);
 
 		// Divide the pixels on its number
 		pixelMean(retImg);
@@ -274,12 +293,11 @@ int* backgroundExtraction(int** &imgArr) {
 
 	// free memory
 	deleteImageArray(buffer, sz);
-	MPI_Type_free(&MPI_IMAGE);
-	MPI_Op_free(&MPI_ADD_IMAGE);
 	delete[] sendImg;
-
-	// Finalize MPI env
-	MPI_Finalize();
-
+	if (wRank != rootProcessor) {
+		delete[] retImg;
+		retImg = nullptr;
+	}
+	
 	return retImg;
 }
